Port CHRoot Builder and rename hashi* imports
====
EOD WIP: I've ported all the client code over, on Monday I will test that it works in a VM before porting the unit tests as the mocking strategy used is heavily reliant on auto rest, and I want to port it to stretchify mocks that don't care as much about the autorest specifics

Don't use Subscription ID as Snapshot URL

Large commit: Chroot unit tests passing, azure-sdk-for-go no longer imported by go.mod, remove references to device code, clear up missing pieces of DTL builder

Fix linux disk attatcher

Remove unnused mock that I added earlier

Re-add Fetching of Tenant ID

Forgot to use New Steps

null check was wrong way for SDK model

see above

Don't try and get tenant ID if we're using CLI auth

Address linter, fix linux test

StepCreateNewDiskTest -> Return proper autorest future fake

remove device code acceptance tests, handle flakey unit test by sorting resulting objects to insure consistant comparison

Update builder/azure/chroot/step_verify_source_disk_test.go


Remove CHRoot Acceptance Test && rename all hashi* imports removing prefix

Make generate and fix lint

Readd subscription, tenant ID fetching for CLI auth, so we know which subscription to save stuff to, this caused acceptance test failures which this commit fixes

Lint

+++ builder/azure/chroot/shared_image_gallery_destination.go
@@ -29,7 +29,7 @@ type TargetRegion struct {
 	// Name of the Azure region
 	Name string `mapstructure:"name" required:"true"`
 	// Number of replicas in this region. Default: 1
-	ReplicaCount int32 `mapstructure:"replicas"`
+	ReplicaCount int64 `mapstructure:"replicas"`
 	// Storage account type: Standard_LRS or Standard_ZRS. Default: Standard_ZRS
 	StorageAccountType string `mapstructure:"storage_account_type"`
 }
